+++ js/jquery/components/OffCanvas.js
@@ -24,6 +24,8 @@ Toolkit.OffCanvas = Toolkit.Component.extend(function(element, options) {
 
     // Setup container
     this.container = element.parents('.' + vendor + 'canvas').addClass(animation);
+    this.primary = element.siblings('.' + vendor + 'on-canvas');
+    this.secondary = element.siblings('.' + vendor + 'off-canvas');
 
     // Determine the side
     this.side = element.hasClass(vendor + 'off-canvas--left') ? 'left' : 'right';
@@ -54,16 +56,16 @@ Toolkit.OffCanvas = Toolkit.Component.extend(function(element, options) {
      * Hide the sidebar and reset the container.
      */
     hide: function() {
-        this.container
-            .removeClass('move-' + this.opposite);
+        this.container.removeClass('move-' + this.opposite);
 
         this.element
             .conceal()
             .removeClass('is-expanded')
-            .aria({
-                hidden: true,
-                expanded: false
-            });
+            .aria('expanded', false);
+
+        if (this.options.stopScroll) {
+            $('body').removeClass('no-scroll');
+        }
 
         this.fireEvent('hide');
     },
@@ -73,27 +75,28 @@ Toolkit.OffCanvas = Toolkit.Component.extend(function(element, options) {
      * If hideOthers is true, hide other open sidebars.
      */
     show: function() {
-        var options = this.options,
-            element = this.element,
-            container = this.container;
+        var options = this.options;
 
         if (options.hideOthers) {
-            container.find('.' + vendor + 'off-canvas').each(function() {
+            this.secondary.each(function() {
                 var sidebar = $(this);
 
-                if (!sidebar.is(element) && sidebar.hasClass('is-expanded')) {
+                if (sidebar.hasClass('is-expanded')) {
                     sidebar.toolkit('offCanvas', 'hide');
                 }
             });
         }
 
-        container
-            .addClass('move-' + this.opposite);
+        this.container.addClass('move-' + this.opposite);
 
-        element
+        this.element
             .reveal()
             .addClass('is-expanded')
             .aria('expanded', true);
+
+        if (options.stopScroll) {
+            $('body').addClass('no-scroll');
+        }
 
         this.fireEvent('show');
     },
@@ -122,7 +125,7 @@ Toolkit.OffCanvas = Toolkit.Component.extend(function(element, options) {
         }
 
         var sidebar = this.element,
-            inner = this.container.find('.' + vendor + 'on-canvas'),
+            inner = this.primary,
             transClass = 'no-transition';
 
         sidebar.addClass(transClass);
@@ -169,7 +172,8 @@ Toolkit.OffCanvas = Toolkit.Component.extend(function(element, options) {
     selector: '',
     animation: 'push',
     openOnLoad: false,
-    hideOthers: true
+    hideOthers: true,
+    stopScroll: true
 });
 
 Toolkit.create('offCanvas', function(options) {