+++ js/jquery/components/Toast.js
@@ -65,11 +65,9 @@ Toolkit.Toast = Toolkit.Component.extend(function(element, options) {
 
         this.fireEvent('hide', element); // Must be called first since the element gets removed
 
-        element
-            .one(Toolkit.transitionEnd, function() {
-                $(this).remove();
-            })
-            .conceal();
+        element.one(Toolkit.transitionEnd, function() {
+            $(this).remove();
+        }).conceal();
     },
 
     /**